examples/nanogpt/requirements.txt
transformers==4.37.2
tiktoken==0.5.2
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.1+cu121
--extra-index-url https://pypi.nvidia.com
nvidia-ammo==0.9.1
datasets==2.18.0
